finaly get the way params works
