+++ src/main.js
@@ -1,7 +1,7 @@
 import Vue from 'vue'
 import App from './App.vue'
 import store from './store'
-import router from './router'
+import router from './router/router'
 
 Vue.config.productionTip = false
