+++ src/store/index.js
@@ -1,6 +1,5 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
-import addExpence from './modules/addExpence'
 import expences from './modules/expences'
 
 
@@ -9,6 +8,5 @@ Vue.use(Vuex)
 export default new Vuex.Store({
   modules: {
     expences,
-    addExpence
   }
 })
