--- src/store/modules/addExpence.js
@@ -1,12 +1,0 @@
-export default {
-  state: {
-
-  },
-  getters: {
-
-  },
-  mutations: {
-  },
-  actions: {
-  },
-}
